Test: Fixed RootView Not Found Error
Since you are running an instrumented test, the Android framework
requires the Dispatchers.Main thread for actually interacting with
the UI since it is the Dispatchers.Main thread that starts the app
before any testing begins. So since the Dispatchers.Main creates the
rootView, any thread that you set as main after cannot access it.
Since recyclerView is attached to the rootView, the recyclerView cannot
 find the rootView of which to attach(?)